requirement.txt
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.2.0
PySide6==6.6.1